Revert checks for all listed Kafka brokers
This shouldn't be necessary with a set replication factor. We observed
issues with the Producer repeatedly attempting to send to a broker that had gone
down while others were available, though this is likely due to the
replication factor of the broker's auto-created topics being set to 1.

We should either look into automatically creating topics with our desired configuration or leave this to the team setting up Kafka.