Remove labels from Prometheus metrics
Counters aren't initialized with the labels, leading to funky behavior
when attempting to aggregate the Counters. It also doesn't really make
sense to increase the cardinality of the metrics by introducing labels
for possible SC/SEC combinations - try to keep the metrics simple.
Consult the logs for failing SC/SECs if the aggregated counters
indicate failures.